requirements.txt
accelerate==0.19.0
datasets==2.12.0
deepspeed==0.9.3
diffusers==0.17.0
einops==0.6.1
ftfy==6.1.1
gradio==3.44.0
imageio==2.31.1
iopath==0.1.10
ipdb==0.13.13
joblib==1.3.1
matplotlib==3.7.1
mdtex2html==1.2.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
peft==0.3.0
Pillow==9.5.0
pytorchvideo==0.1.5
PyYAML==6.0
PyYAML==6.0.1
regex==2023.6.3
scipy==1.11.2
timm==0.9.2
torch==1.13.1
torchaudio==0.13.1
torchaudio==0.13.1+cu116
torchvision==0.14.1
torchvision==0.14.1+cu116
tqdm==4.65.0
transformers==4.29.2